refactor: タブの表示内容を分離し、URLオープンダイアログの処理を整理
